requirements.txt
# ---------- requirements.txt ----------
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
passlib[bcrypt]
pyotp
python-jose[cryptography]
pydantic
httpx
email-validator
openai